refactoring: move to monitoring core decoupled monitoring resources
