+++ addons/monitoring-addon/monitoring-core/src/main/java/org/kie/kogito/monitoring/core/system/interceptor/MetricsInterceptor.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.monitoring.system.interceptor;
+package org.kie.kogito.monitoring.core.system.interceptor;
 
 import java.util.List;
 
@@ -21,7 +21,7 @@ import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerResponseContext;
 import javax.ws.rs.container.ContainerResponseFilter;
 
-import org.kie.kogito.monitoring.system.metrics.SystemMetricsCollector;
+import org.kie.kogito.monitoring.core.system.metrics.SystemMetricsCollector;
 
 public class MetricsInterceptor implements ContainerResponseFilter {
