+++ addons/monitoring-addon/monitoring-core/src/main/java/org/kie/kogito/monitoring/core/system/metrics/dmnhandlers/DecisionConstants.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.monitoring.system.metrics.dmnhandlers;
+package org.kie.kogito.monitoring.core.system.metrics.dmnhandlers;
 
 public class DecisionConstants {
