+++ kogito-bom/pom.xml
@@ -126,6 +126,19 @@
 
       <dependency>
         <groupId>org.kie.kogito</groupId>
+        <artifactId>monitoring-core</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.kie.kogito</groupId>
+        <artifactId>monitoring-core</artifactId>
+        <version>${project.version}</version>
+        <classifier>sources</classifier>
+      </dependency>
+
+
+      <dependency>
+        <groupId>org.kie.kogito</groupId>
         <artifactId>grafana-api</artifactId>
         <version>${project.version}</version>
       </dependency>
